Save last changes

+++ src/main/java/io/javabrains/weaponservice/controller/WeaponController.java
@@ -1,8 +1,10 @@
 package io.javabrains.weaponservice.controller;
 
+import javax.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import io.javabrains.weaponservice.weapon.WeaponResponse;
 import io.javabrains.weaponservice.model.Weapon;
@@ -33,38 +36,59 @@ public class WeaponController {
   Logger logger = LoggerFactory.getLogger(WeaponController.class);
 
   @RequestMapping(method = RequestMethod.GET)
-  public WeaponResponse weapons() {
+  public WeaponResponse weapons(HttpServletRequest request) {
+    if (!weaponService.isTokenValidBossOrCreator(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     return weaponService.getAllWeapons();
   }
 
   @GetMapping("/{Id}")
-  public ResponseEntity<?> getWeaponById(@PathVariable("Id") Long Id) {
+  public ResponseEntity<?> getWeaponById(@PathVariable("Id") Long Id, HttpServletRequest request) {
+    if (!weaponService.isTokenValidBossOrCreator(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     return weaponService.getById(Id);
   }
 
   @RequestMapping(method = RequestMethod.POST)
-  public Weapon addweapon (@RequestBody Weapon weapon){
+  public Weapon addweapon (@RequestBody Weapon weapon, HttpServletRequest request){
+    if (!weaponService.isTokenValidBoss(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     logger.info("Saving weapon: {}", weapon);
     return weaponService.create(weapon);
   }
   
   @PatchMapping("/{Id}")
-  public ResponseEntity<?> updateWeapon(@PathVariable("Id") Long Id, @RequestBody Weapon weapon){
+  public ResponseEntity<?> updateWeapon(@PathVariable("Id") Long Id, @RequestBody Weapon weapon, HttpServletRequest request) {
+    if (!weaponService.isTokenValidCreator(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     return weaponService.updateById(Id, weapon);
   }
   
   @PatchMapping("/{Id}/addBand")
-  public ResponseEntity<Object> updateWeaponsBand(@PathVariable("Id") Long Id, @RequestBody String bandName) {
+  public ResponseEntity<Object> updateWeaponsBand(@PathVariable("Id") Long Id, @RequestBody String bandName, HttpServletRequest request) {
+    if (!weaponService.isTokenValidBoss(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     return weaponService.addBand(Id, bandName);
   }
 
   @PatchMapping("/{Id}/addTask")
-  public ResponseEntity<Object> updateWeaponsTask(@PathVariable("Id") Long Id, @RequestBody String taskName) {
+  public ResponseEntity<Object> updateWeaponsTask(@PathVariable("Id") Long Id, @RequestBody String taskName, HttpServletRequest request) {
+    if (!weaponService.isTokenValidBoss(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     return weaponService.addTask(Id, taskName);
   }
 
   @RequestMapping(value="/{Id}", method = RequestMethod.DELETE)
-  public ResponseEntity<?> deleteWeapon (@PathVariable("Id") Long Id) {
+  public ResponseEntity<?> deleteWeapon (@PathVariable("Id") Long Id, HttpServletRequest request) {
+    if (!weaponService.isTokenValidCreator(request)) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    }
     logger.info("Deleting the weapon with and id {}", Id);
     return weaponService.delete(Id);
   }
